allow custom_ipxe to handle both paths for booting an ipxe script
