+++ installers/custom_ipxe/ipxe_script_test.go
@@ -30,7 +30,7 @@ func TestScript(t *testing.T) {
 			s.Set("tinkerbell", "http://127.0.0.1")
 			s.Set("ipxe_cloud_config", "packet")
 
-			bootScript(m.Job(), &s)
+			ipxeScript(m.Job(), &s)
 			got := string(s.Bytes())
 			if script != got {
 				t.Fatalf("%s bad iPXE script:\n%v", typ, diff.LineDiff(script, got))
